+++ HISTORY.rst
@@ -36,6 +36,7 @@ History
     * Card #273: Added new state to Assignment: post_processing (rudaporto).
     * Card #273: Added new transitions to move to and back in_qa to post_processing and to approve from post_processing (rudaporto).
     * Reclassify Report views to be marked as background tasks in newrelic agent (rudaporto).
+    * Card #286: Added remove_confirmation transition to LeadOrder workflow (rudaporto).
 
 2.1.12 (2017-04-28)
 -------------------